requirements.txt
python-dotenv~=1.0.0
crewai~=0.28.6
langchain~=0.1.11
replicate~=0.24.0